feat: ISY-655 add IsyServiceApiSicherheitAutoConfiguration
